style(preferences): hooks
